+++ lib/declaration.ml
@@ -1277,6 +1277,8 @@ let read_type_value : type a. a property -> Cursor.t -> declaration option =
   | White_space -> Some (v White_space (read_white_space t))
   | White_space_collapse ->
       Some (v White_space_collapse (read_white_space_collapse t))
+  | Font_variant_alternates ->
+      Some (v Font_variant_alternates (read_font_variant_alternates t))
   | Text_decoration -> Some (v Text_decoration (read_text_decoration t))
   | Text_decoration_line ->
       Some (v Text_decoration_line (read_text_decoration_lines t))